requirements.txt
numpy>=1.26
pandas>=2.1
PyYAML>=6.0
scipy>=1.11
shapely>=2.0
pyproj>=3.6
rasterio>=1.3
xarray>=2024.1
rioxarray>=0.15
matplotlib>=3.8
pytest>=8.0